# contracts/knowledge-public.openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OrganisationAuditEventResponse": {
        "properties": {
          "action": {
            "title": "Action",
            "type": "string"
          },
          "actor_id": {
            "title": "Actor Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "subject_id": {
            "title": "Subject Id",
            "type": "string"
          }
        },
        "required": [
          "actor_id",
          "action",
          "subject_id",
          "created_at"
        ],
        "title": "OrganisationAuditEventResponse",
        "type": "object"
      },
      "OrganisationResponse": {
        "properties": {
          "people": {
            "items": {
              "$ref": "#/components/schemas/PersonSchema"
            },
            "title": "People",
            "type": "array"
          },
          "products": {
            "items": {
              "$ref": "#/components/schemas/ProductSchema"
            },
            "title": "Products",
            "type": "array"
          },
          "squads": {
            "items": {
              "$ref": "#/components/schemas/SquadSchema"
            },
            "title": "Squads",
            "type": "array"
          },
          "value_streams": {
            "items": {
              "$ref": "#/components/schemas/ValueStreamSchema"
            },
            "title": "Value Streams",
            "type": "array"
          }
        },
        "required": [
          "people",
          "value_streams",
          "products",
          "squads"
        ],
        "title": "OrganisationResponse",
        "type": "object"
      },
...
      "PersonRequest": {
        "description": "``expected_revision`` is omitted to create and required to update.",
        "properties": {
          "expected_revision": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Revision"
          },
          "person": {
            "$ref": "#/components/schemas/PersonSchema"
          }
        },
        "required": [
          "person"
        ],
        "title": "PersonRequest",
        "type": "object"
      },
...
      "ProductRequest": {
        "properties": {
          "expected_revision": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Revision"
          },
          "product": {
            "$ref": "#/components/schemas/ProductSchema"
          }
        },
        "required": [
          "product"
        ],
        "title": "ProductRequest",
        "type": "object"
      },
...
      "RemovalRequest": {
        "properties": {
          "expected_revision": {
            "title": "Expected Revision",
            "type": "integer"
          }
        },
        "required": [
          "expected_revision"
        ],
        "title": "RemovalRequest",
        "type": "object"
      },
...
      "SquadRequest": {
        "properties": {
          "expected_revision": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Revision"
          },
          "squad": {
            "$ref": "#/components/schemas/SquadSchema"
          }
        },
        "required": [
          "squad"
        ],
        "title": "SquadRequest",
        "type": "object"
      },
...
      "SystemOwnershipResponse": {
        "properties": {
          "products": {
            "items": {
              "$ref": "#/components/schemas/ProductSchema"
            },
            "title": "Products",
            "type": "array"
          },
          "squads": {
            "items": {
              "$ref": "#/components/schemas/SquadSchema"
            },
            "title": "Squads",
            "type": "array"
          },
          "system_id": {
            "title": "System Id",
            "type": "string"
          },
          "value_streams": {
            "items": {
              "$ref": "#/components/schemas/ValueStreamSchema"
            },
            "title": "Value Streams",
            "type": "array"
          }
        },
        "required": [
          "system_id",
          "squads",
          "products",
          "value_streams"
        ],
        "title": "SystemOwnershipResponse",
        "type": "object"
      },
...
      "ValueStreamRequest": {
        "properties": {
          "expected_revision": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Revision"
          },
          "value_stream": {
            "$ref": "#/components/schemas/ValueStreamSchema"
          }
        },
        "required": [
          "value_stream"
        ],
        "title": "ValueStreamRequest",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Knowledge portal",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/organisation": {
      "get": {
        "operationId": "view_organisation_get",
        "parameters": [
          {
            "in": "header",
            "name": "X-Fake-Actor-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Fake-Actor-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganisationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "View",
        "tags": [
          "organisation"
        ]
      }
    },
    "/organisation/audit": {
      "get": {
        "operationId": "audit_organisation_audit_get",
        "parameters": [
          {
            "in": "header",
            "name": "X-Fake-Actor-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Fake-Actor-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/OrganisationAuditEventResponse"
                  },
                  "title": "Response Audit Organisation Audit Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Audit",
        "tags": [
          "organisation"
        ]
      }
    },
    "/organisation/people": {
      "post": {
        "operationId": "create_person_organisation_people_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Fake-Actor-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Fake-Actor-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PersonRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganisationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Person",
        "tags": [
          "organisation"
        ]
      }
    },
    "/organisation/people/{person_id}": {
      "put": {
        "operationId": "update_person_organisation_people__person_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "person_id",
            "required": true,
            "schema": {
              "title": "Person Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Fake-Actor-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Fake-Actor-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PersonRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganisationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Person",
        "tags": [
          "organisation"
        ]
      }
    },
    "/organisation/products": {
      "post": {
        "operationId": "create_product_organisation_products_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Fake-Actor-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Fake-Actor-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganisationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Product",
        "tags": [
          "organisation"
        ]
      }
    },
    "/organisation/products/{product_id}": {
      "delete": {
        "operationId": "remove_product_organisation_products__product_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "product_id",
            "required": true,
            "schema": {
              "title": "Product Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Fake-Actor-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Fake-Actor-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RemovalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganisationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Remove Product",
        "tags": [
          "organisation"
        ]
      },
      "put": {
        "operationId": "update_product_organisation_products__product_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "product_id",
            "required": true,
            "schema": {
              "title": "Product Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Fake-Actor-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Fake-Actor-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganisationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Product",
        "tags": [
          "organisation"
        ]
      }
    },
    "/organisation/squads": {
      "post": {
        "operationId": "create_squad_organisation_squads_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Fake-Actor-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Fake-Actor-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SquadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganisationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Squad",
        "tags": [
          "organisation"
        ]
      }
    },
    "/organisation/squads/{squad_id}": {
      "delete": {
        "operationId": "remove_squad_organisation_squads__squad_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "squad_id",
            "required": true,
            "schema": {
              "title": "Squad Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Fake-Actor-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Fake-Actor-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RemovalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganisationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Remove Squad",
        "tags": [
          "organisation"
        ]
      },
      "put": {
        "operationId": "update_squad_organisation_squads__squad_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "squad_id",
            "required": true,
            "schema": {
              "title": "Squad Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Fake-Actor-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Fake-Actor-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SquadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganisationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Squad",
        "tags": [
          "organisation"
        ]
      }
    },
    "/organisation/systems/{system_id}/ownership": {
      "get": {
        "operationId": "system_ownership_organisation_systems__system_id__ownership_get",
        "parameters": [
          {
            "in": "path",
            "name": "system_id",
            "required": true,
            "schema": {
              "title": "System Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Fake-Actor-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Fake-Actor-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemOwnershipResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "System Ownership",
        "tags": [
          "organisation"
        ]
      }
    },
    "/organisation/value-streams": {
      "post": {
        "operationId": "create_value_stream_organisation_value_streams_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Fake-Actor-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Fake-Actor-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValueStreamRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganisationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Value Stream",
        "tags": [
          "organisation"
        ]
      }
    },
    "/organisation/value-streams/{value_stream_id}": {
      "delete": {
        "operationId": "remove_value_stream_organisation_value_streams__value_stream_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "value_stream_id",
            "required": true,
            "schema": {
              "title": "Value Stream Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Fake-Actor-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Fake-Actor-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RemovalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganisationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Remove Value Stream",
        "tags": [
          "organisation"
        ]
      },
      "put": {
        "operationId": "update_value_stream_organisation_value_streams__value_stream_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "value_stream_id",
            "required": true,
            "schema": {
              "title": "Value Stream Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Fake-Actor-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Fake-Actor-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValueStreamRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganisationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Value Stream",
        "tags": [
          "organisation"
        ]
      }
    },
...
  }
}